Added .editorconfig for consistent coding style
This commit introduces a new .editorconfig file to the project. It sets some basic rules for the code editor, such as using UTF-8 charset, Unix-style line endings, spaces for indentation with an indent size of 2. It also ensures that files end with a newline and trailing whitespaces are trimmed. Exceptions are made for .md and .enc files where trailing whitespaces aren't trimmed and final newline isn't inserted respectively.