Add ChatPage component
A new ChatPage component has been added to the client/src/components directory. This component is a functional React component that receives a 'socket' prop. Currently, it returns an empty div with the class name "chat". The purpose of this commit is to set up the basic structure for future development of chat functionality in our application.