Implement user tracking and broadcast on server
Added functionality to track connected users in the server. A new array 'users' has been introduced to store the details of all connected users. On a new connection, the 'newUser' event is listened for, which pushes the new user's data into this array and emits a 'newUserResponse' with updated users list. Also, when a user disconnects, their data is removed from the 'users' array and an updated list is broadcasted again. This helps in keeping all clients updated about active users at any given time.