Add typing status to chat components
Added a 'typingStatus' prop to the ChatBody component and updated the message status display to reflect this new prop. Also, in the ChatFooter component, implemented a 'handleTyping' function that emits a 'typing' event with the user's name when they are typing. This function is triggered on keydown in the message input field. After sending a message, another 'typing' event is emitted with an empty string to reset the typing status.